src/rag/requirements.txt
# Core RAG dependencies
llama-index==0.10.0
chromadb==0.4.22
python-dotenv==1.0.0

# LLM providers - LlamaIndex integrations
llama-index-llms-ollama==0.1.3
llama-index-llms-openai==0.1.5
llama-index-embeddings-huggingface==0.1.4
openai==1.12.0
ollama==0.1.6

# Data processing
pypdf==4.0.1
docx2txt==0.8
python-pptx==0.6.23
openpyxl==3.1.2
pandas==2.2.0

# Utilities
tqdm==4.66.1
colorama==0.4.6

# API server
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
